+++ src/Gateway/HTTP/SimpleHTTPClient.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const http = require('http');
+const https = require('https');
 const axios = require('axios');
 const Response = require('./Response/Response');
 const InvalidArgumentException = require('../../Error/InvalidArgumentException');
@@ -12,6 +14,16 @@ class SimpleHTTPClient {
      */
     constructor(configuration) {
         this.configuration = configuration;
+
+        const agentOptions = { keepAlive: this.configuration.keepAlive };
+
+        this._httpAgent = new http.Agent(agentOptions);
+        this._httpsAgent = new https.Agent(agentOptions);
+
+        this._axiosInstance = axios.create({
+            httpAgent: this._httpAgent,
+            httpsAgent: this._httpsAgent
+        });
     }
 
     get configuration() {
@@ -26,6 +38,19 @@ class SimpleHTTPClient {
             throw new InvalidArgumentException('Configuration should be a Configuration object');
         } else {
             this._configuration = configuration;
+        }
+    }
+
+    /**
+     * Destroys underlying HTTP/HTTPS agents, closing all open connections
+     **/
+    destroy() {
+        if (this._httpAgent) {
+            this._httpAgent.destroy();
+        }
+
+        if (this._httpsAgent) {
+            this._httpsAgent.destroy();
         }
     }
 
@@ -107,7 +132,7 @@ class SimpleHTTPClient {
         }
 
         try {
-            const rawResponse = await axios.request(requestOptions);
+            const rawResponse = await this._axiosInstance.request(requestOptions);
 
             return new Response(rawResponse.data, rawResponse.status, {
                 'Content-Type': 'application/json; encoding=UTF-8'